Create `CrewManager` component
