+++ dragon-boat/components/CrewManager/CrewManager.scss
@@ -1,0 +1,31 @@
+@import url(../../colors.css);
+
+.crew-manager {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: center;
+	gap: 36px;
+
+	@media (max-width: 768px) {
+		flex-direction: column;
+		gap: 12px;
+	}
+
+	.main {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		gap: 12px;
+
+		.button-group {
+			display: flex;
+			gap: 4px;
+			width: 100%;
+
+			button {
+				flex: 1;
+			}
+		}
+	}
+}
